proj3/requirements.txt
crewai
python-dotenv
langchain-openai